Fix #67: enable `system-ghc: true`. (#83)
Also changes `Pier.Core.Command` so that absolute paths of "external" files
aren't hashed.  This is a little of a workaround due to the system ghc's libdir
being a directory.  But doing a deep hash of the system GHC seems like overkill
(and is also unnecessary on, e.g., Nix).  We also do get some recompile
checking from Shake around the use of the `PATH` env variable (though I'm not
exactly sure how much).